refactor: implement Linux camera capture (V4L2+PipeWire) and fix clang-tidy warnings
- Add V4L2 and PipeWire dual-backend camera capture for Linux
  - V4L2: mmap streaming, YUYV/NV12 → RGB24 via libyuv, fallback
  - PipeWire: pw_thread_loop + pw_stream with SPA POD format negotiation
  - Factory function create_capture_device() with PW→V4L2→null fallback
  - Camera enumeration via /dev/video* + VIDIOC_QUERYCAP
- Switch main.cpp to use factory pattern
- Add libpipewire-0.3 system dependency in xmake.lua
- Fix .clang-tidy: strip noisy checks, enable meaningful bugprone/performance/modernize
- Fix .clang-format: ColumnLimit 80→120
- Fix all clang-tidy warnings across camera/pipeline/socket/socket/messages/seetaface
  - Narrow enum base types to std::uint8_t
  - Add [[nodiscard]] and explicit, Rule-of-5 consistency
  - Replace std::call_once with lambda IIFE (IDE LSP compat)
  - Use std::unique_ptr with release pattern for SeetaImageData allocation
  - Remove redundant std::move on trivially-copyable types
  - Uppercase float literal suffixes (0.0f→0.0F)
- Functional style: lambda helpers, pure helpers, const correctness

+++ src/facerecognizer/protocol/messages.h
@@ -8,7 +8,7 @@
 
 namespace su::facerecognizer::protocol {
 
-enum class CommandType : std::uint16_t {
+enum class CommandType : std::uint8_t {
     kUnknown = 0,
     kStartCameraPreview = 1,
     kStopCameraPreview = 2,
@@ -18,7 +18,7 @@ enum class CommandType : std::uint16_t {
     kCompareFeatures = 6,
 };
 
-enum class MessageType : std::uint16_t {
+enum class MessageType : std::uint8_t {
     kError = 0,
     kPreviewFrame = 1,
     kCaptureResult = 2,
@@ -39,8 +39,8 @@ struct CommandMessage {
     int preview_fps = 15;
     bool extract_feature = false;
     bool liveness_detection = false;
-    float face_threshold = 0.62f;
-    float liveness_threshold = 0.80f;
+    float face_threshold = 0.62F;
+    float liveness_threshold = 0.80F;
     std::string username_hint;
     std::vector<float> feature_a;
     std::vector<float> feature_b;
@@ -66,14 +66,14 @@ struct CaptureResultMessage {
 struct RecognitionResultMessage {
     bool matched = false;
     std::string username;
-    float similarity = 0.0f;
+    float similarity = 0.0F;
     std::optional<BoundingBox> face_box;
     std::vector<float> feature;
 };
 
 struct CompareResultMessage {
     bool ok = false;
-    float similarity = 0.0f;
+    float similarity = 0.0F;
     std::string error_message;
 };
